refactor: 애플리케이션 구조로 경로 수정
- 클린 아키텍처 참조 규칙 준수
