+++ src/application/repository.ts
@@ -1,4 +1,4 @@
-import NotificationEntity from "./entity";
+import NotificationEntity from "src/domain/entity";
 
 export default interface INotificationRepository {
     save(entity: NotificationEntity): Promise<NotificationEntity>;
