+++ src/application/service.ts
@@ -1,11 +1,13 @@
 import { Inject, Injectable } from "@nestjs/common";
 
 import NotificationEntity from "src/domain/entity";
-import INotificationRepository from "src/domain/repository";
-import { CreateRequestDTO } from "src/presentation/dto";
+
+import INotificationRepository from "src/application/repository";
+
+import { CreateRequestDTO } from "src/presentation/dto"; // 역참조
 
 @Injectable()
-export class NotificationService {
+export default class NotificationService {
     constructor(
         @Inject("INotificationRepository")
         private readonly repository: INotificationRepository,
